Rename availability service
