+++ web/src/services/availability.ts
